% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distVincentyEllipsoid.R
\name{distVincentyEllipsoid}
\alias{distVincentyEllipsoid}
\title{'Vincenty' (ellipsoid) great circle distance}
\usage{
distVincentyEllipsoid(p1, p2, a = 6378137, b = 6356752.3142,
  f = 1/298.257223563)
}
\arguments{
\item{p1}{longitude/latitude of point(s), in degrees 1; can be a vector of
two numbers, a matrix of 2 columns (first one is longitude, second is
latitude) or a SpatialPoints* object}

\item{p2}{as above}

\item{a}{Equatorial axis of ellipsoid}

\item{b}{Polar axis of ellipsoid}

\item{f}{Inverse flattening of ellipsoid}
}
\value{
Distance value in the same units as the ellipsoid (default is meters)
}
\description{
The shortest distance between two points (i.e., the 'great-circle-distance'
or 'as the crow flies'), according to the 'Vincenty (ellipsoid)' method. 
This method uses an ellipsoid and the results are very accurate. 
The method is computationally more intensive than the other great-circled methods in this package.
}
\section{Source}{

Function borrowed to the 'geosphere' package to attenuate packages' dependency.
See this package and the function help for more details.
}

\author{
Chris Veness and Robert Hijmans
}
