% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xytb-class.R
\name{dxyt2}
\alias{dxyt2}
\title{internal function}
\usage{
dxyt2(dxyt, winsize = seq(3, 13, 2), idquant = seq(0, 1, 0.25))
}
\arguments{
\item{dxyt}{A parameter}

\item{winsize}{A parameter}

\item{idquant}{A parameter}
}
\description{
internal function
}
\author{
Laurent Dubroca
}
