% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xytb-class.R
\name{ltraj2xytb}
\alias{ltraj2xytb}
\title{ltraj object conversion to xytb object}
\usage{
ltraj2xytb(ltraj)
}
\arguments{
\item{ltraj:}{a ltraj object}
}
\value{
A xytb object with behavioural information taken from the infolocs
slots by default.
}
\description{
The function converts a ltraj object in an  xytb object 
(see \link{xytb-class}).
}
\examples{
\dontrun{ 
#a dataset:
str(track_CAGA_005)
xytb<-xytb(track_CAGA_005,"a track",3,.5)
ltraj<-xytb2ltraj(xytb)
#all adehabitatLT function are now available
summary(ltraj)
plot(ltraj)
}
}

