% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xytb-class.R
\name{modelRF}
\alias{modelRF}
\title{xytb randomForest function}
\usage{
modelRF(xytb, type = c("actual", "shifted"), nob = "-1", colin = T,
  varkeep = c("v", "dist", "thetarel"), zerovar = T, ntree = 501,
  importance = T, ...)
}
\arguments{
\item{xytb:}{an xytb object}

\item{type:}{character -actual or shifted- use actual data or shifted one to
build the model}

\item{nob:}{character. Define the unobserved value of the behaviour (and
where prediction are done)}

\item{colin:}{boolean - remove colinearity among predictors (see the caret
package for more details)}

\item{varkeep:}{character vector - the variables names in this vector are
keeped in the model even if colinearity is found (usefull to keep 'classical'
parameters and to help interpretation)}

\item{zerovar:}{boolean - remove near zero variance predictor (see the caret
package for more details)}

\item{ntree:}{number of trees in the random Forest (see the randomForest
package for more details)}

\item{importance:}{boolean (see the randomForest package for more details)}
}
\description{
Build a random forest model on an xytb object, predicting behaviour using
regular observation (type actual) or shifted one (type shifted). Parameters
are passed to the randomForest function if needed.
}
\author{
Laurent Dubroca and Andréa Thiébault
}

