% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xytb-class.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,xytb,missing-method}
\title{xytb plot method}
\usage{
\S4method{plot}{xytb,missing}(x, y)
}
\arguments{
\item{x}{An xytb object.}

\item{y}{empty}
}
\value{
a plot
}
\description{
plot method for xytb object
}
