% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xytb-class.R
\name{run}
\alias{run}
\title{#compute a random forest model to predict behaviour (b, where -1 is
#unobserved behaviour) using the derived
#parameters ("actual")
#extract the model
# results from randomForest package:}
\usage{
run()
}
\description{
#compute a random forest model to predict behaviour (b, where -1 is
#unobserved behaviour) using the derived
#parameters ("actual")
#extract the model
# results from randomForest package:
}
