% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xytb-class.R
\name{shiftvalue}
\alias{shiftvalue}
\title{internal function}
\usage{
shiftvalue(dat0, mov = seq(5, 250, 5))
}
\arguments{
\item{dat0}{A parameter}

\item{mov}{A parameter}
}
\description{
internal function
}
\author{
Laurent Dubroca
}
