% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xytb-class.R
\name{test}
\alias{test}
\title{internal test function for dev purposes}
\usage{
test()
}
\description{
blablablablbalalb
}
\keyword{#track_CAGA_005}
\keyword{a}
\keyword{dataset}
\keyword{internal}
\keyword{is}
