% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xytb-class.R
\docType{class}
\name{xytb-class}
\alias{xytb-class}
\title{xytb class definition}
\description{
xytb is an trajectory object with observed behaviour
}
\section{Slots}{

\describe{
\item{\code{desc}}{}

\item{\code{xyt}}{}

\item{\code{b}}{}

\item{\code{dxyt}}{}

\item{\code{befdxyt}}{}

\item{\code{model}}{}

\item{\code{predb}}{}
}}

