% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xytb-class.R
\name{xytb2ltraj}
\alias{xytb2ltraj}
\title{xytb class conversion to ltraj object}
\usage{
xytb2ltraj(xytb)
}
\arguments{
\item{xytb:}{an xytb object}
}
\value{
A ltraj object with behavioural information recorded in the infoloc
}
\description{
The function converts a xytb object in ltraj object as defined in the
adehabitatLT package (see \link[adehabitatLT]{as.ltraj}).
}
\examples{
\dontrun{ 
#a dataset:
str(track_CAGA_005)
xytb<-xytb(track_CAGA_005,"a track",3,.5)
ltraj<-xytb2ltraj(xytb)
#all adehabitatLT function are now available
summary(ltraj)
plot(ltraj)
}
}

