% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xytb-class.R
\name{ltraj2xytb}
\alias{ltraj2xytb}
\title{ltraj object conversion to xytb object}
\usage{
ltraj2xytb(ltraj, desc = "ltraj object convert to xytb")
}
\arguments{
\item{ltraj}{A ltraj object}

\item{desc}{General descriptor of the data}
}
\value{
A xytb object with behavioural information taken from the infolocs
(aka ltraj object should correspond the an export of xytb object in ltraj).
}
\description{
The function converts a ltraj object to a xytb object 
(see  
\url{https://CRAN.R-project.org/package=adehabitatLT}
and \link{xytb-class}).
}
\examples{
#track_CAGA_005 is a dataset
xytb<-xytb(track_CAGA_005,"a track",3,.5)
ltraj<-xytb2ltraj(xytb)
xytb2<-ltraj2xytb(ltraj,"a track")
}
\author{
Laurent Dubroca
}
