% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xytb-class.R
\name{xytb2ltraj}
\alias{xytb2ltraj}
\title{xytb class conversion to ltraj object}
\usage{
xytb2ltraj(xytb)
}
\arguments{
\item{xytb}{An xytb object}
}
\value{
A ltraj object with behavioural information recorded in the infoloc
}
\description{
The function converts a xytb object to a ltraj object as defined in the
adehabitatLT package (see 
\url{https://CRAN.R-project.org/package=adehabitatLT}
\link[adehabitatLT]{as.ltraj}).
}
\examples{
#track_CAGA_005 is a dataset
xytb<-xytb(track_CAGA_005,"a track",3,.5)
ltraj<-xytb2ltraj(xytb)
#all adehabitatLT function are now available
summary(ltraj)
plot(ltraj)
}
\author{
Laurent Dubroca
}
